Add identity `convert`
When wrapping a `@bitfield` or `@bitflags` struct in a struct with a
type parameter, julia inserts calls to `convert` arguments to the
corresponding field type. If the object being passed in is already
of the correct type, this should be a noop, but due to a missing
`convert` method, this errored at runtime. This fixes this issue,
by adding the missing conversion method.

Fixes #10.